added logic to circumvent callback on specific methods

+++ aircond_rails_app/app/models/aircond.rb
@@ -48,12 +48,16 @@ class Aircond < ApplicationRecord
 	end
 
 	def update_firebase
+    puts caller[0][/`.*'/][1..-2]
+    return true if caller[0][/`.*'/][1..-2] == "update_website_from_firebase"
 		firebase = Firebase::Client.new("https://nextaircon-6d849.firebaseio.com")
 		data = self.slice(:alias,:temperature,:mode,:fan_speed,:aircond_group_id)
 		firebase.update('/airconds/'+self.id.to_s, data)		
 	end
 
 	def check_state
+    puts caller[0][/`.*'/][1..-2]
+    return true if caller[0][/`.*'/][1..-2] == "update_website_from_firebase"
 		response = send_signal(get_command) if !(self.changes.keys & ["status","temperature","mode","fan_speed"]).empty?
 
     throw :abort if response == false
